Add LOAD_CLASS opcode
Summary:
We don't have a way to put a type object on the stack
unless we already have an instance of it (see `LOAD_TYPE`).

This is needed for static `dataclass`,
since we need to load default values off of specific types
without worrying about subclasses or whether the type is aliased in `globals`.

Introduce a new opcode that loads a type given its `TypeDescr`.

Reviewed By: carljm

Differential Revision: D36315538

fbshipit-source-id: ce5fef0e8d8d33658f47d1e0d9320dc10439d087