Allow explicit defintion of __dict__ and __weakref__ fields
Summary:
Currently in strict modules (and static modules) we can opt-in to allowing a dict by declaring it as a class level member, and we'll generate a slot for it.  But when we do this we generate a `LOAD_FIELD` against it and that doesn't work because the `__dict__` slot isn't a simple slot - there's all sorts of handling caches.

We probably don't actually need to load the `__dict__` from static code very often in the end, but there's no reason to blow up with a class loader exception either.

Reviewed By: sinancepel

Differential Revision: D36293868

fbshipit-source-id: 825259a5aeef4a254ba733112d4161a53c044bb6