Add some more logs
Summary:
These should probably be moved inside the `CallerContext`, but
that uses a forward declaration + a header file implementation (so we
the compiler cannot resolve calls to methods of the `CallerContext.loader`.

Changing this seems to require changes to a bunch of files that include
the `caller_context_impl.h` header. Happy to look into that, but I
lazily decided to put this up for review before :)

Also added some logs to help figure out which file is looked up (and
where), given a module name. This will help us understand whether the
right stubs are being found and loaded, where we're looking for them,
etc.

Reviewed By: DinoV

Differential Revision: D36330840

fbshipit-source-id: a40719da69fc8de31adf03aeecc652240199c39e