Add a non-raising API to clear cached properties
Summary: Adds an API for clearing `cached_property` objects.  This can be simulated with a `__set__` and then a delete against it (to avoid raising an exception on the delete, which would de-optimize the calling method in the JIT), but that's very convoluted way of doing it.  This still ends up being a little weird - you need to access the `cached_property` via the class, and then pass in the instance as the first argument.  But it's a lot better then calling a dunder method.

Reviewed By: sinancepel

Differential Revision: D36319182

fbshipit-source-id: 6292ff6bf9008017026789b4d7133cc7c2e23e39