Add cached_property API to check for value
Summary: Similar to `clear` there's no way to check if a cached property has a value without at best taking an exception (e.g. accessing the underlying slot), and it's kind of a pain to deal with slot vs dict.  This adds an API to see if there is a value.

Reviewed By: carljm

Differential Revision: D36378893

fbshipit-source-id: 0dbb5cd4fa34c3f6aa0c8dbf0287374d66ec0d31