Use string_view in various places in the LIR block builder
Summary:
This is a minor speedup in compile time by doing less allocations, and in the case of GetIdAndType, less string comparisons.
This also removes the use of `std::stringstream` for splitting lines, as there's a significant virtual dispatch cost, and it can't produce string_view's.

This is good for ~1.5% compile time win

Reviewed By: tiansihu

Differential Revision: D36428909

fbshipit-source-id: 960f6e77ef879f83be9a0ee0640f29dabcdbc7a8