remove dead code and confusing comment
Summary:
This `const_stack` and `in_consts` code in the flow graph optimizer was confusing to me,
and when I dug into it, it seems like it's a no-op. We push some things onto a stack, then
later we clear the stack, but we never do anything with the stack. Also not seeing any code
in the C compiler that it corresponds to. And deleting it doesn't make any tests fail.

Also remove a comment that I found confusing, because `if const` is not optimized at the
AST level, it's optimized in the flow graph optimizer.

Reviewed By: sinancepel

Differential Revision: D36464659

fbshipit-source-id: 2a1c5efa2b6070a6ac55f3fce39f3505e9851244