Readonly: Add CHECK_LOAD_ATTR readonly operation
Summary:
Add a new readonly operation `CHECK_LOAD_ATTR`, which is inserted before `LOAD_ATTR` opcode to check if the load is readonly legit.

Currently, the check is carried out conservatively for perf reasons, where as long as there is one descriptor that could violate our readonly rules, we will fire an error. We will consider a more fine-grained solution in the futuer.

Differential Revision: D36287191

fbshipit-source-id: 0f251fa14233ebafdb252ce57163386e86a31941