Fix issue when Simplify creates unreachable loop
Summary:
Whenever Simplify changed a `CondBranch` into a `Branch`, where the
block not taken creates a self-connected loop, the pass would hang
forever.

This was a problem in CFGs where a `CondBranch` branches into a block that
has two predecessors with a `Phi` connecting them, and the
non-`CondBranch` block's input into the `Phi` being the result of that
`Phi`.

Removing the `CondBranch` turns the `Phi` into an `Assign` of the second
predecessor, and since the second predecessor was the output of the
`Phi`, `CopyPropagation` infinitely looped trying to find the original
register of the two `Assign`s.

By making `PhiElimination`s change self-referential `Phi`s into a
`LoadConst<Bottom>` instead of an `Assign`, we ensure that
`CopyPropagation` doesn't enter an infinite loop.

Reviewed By: tekknolagi

Differential Revision: D36082982

fbshipit-source-id: d2889b142499540dd2b725cbf511f06e9f39a8cc