Use correct base address when registering JIT code for perf maps
Summary: The entry point to a function is not necessarily the top of all code. We sometimes put things like alternate entry points, trampolines, etc. before the default vector-call entry. This fix makes `populateCodeSections()` use `orig_entry` which is the value returned by the code memory allocator for all code generated for the current function. Note this is also now consistent with code disassembly on line 1444 in `gen_asm.cpp`.

Reviewed By: mpage

Differential Revision: D36580206

fbshipit-source-id: 0530d229fbc0e5c77a8eeb29dd693d324c4f34ae