Eliminate a bunch of falsey dynamic_casts
Summary:
While `dynamic_cast` with the correct type is fast, as it ends up as a couple of pointer comparisons, it is dramatically slower when the cast is going to fail, as it will first fall back to a bunch of string comparisons against type names before returning the failure.
This reworks things to check the correctness of the dynamic_cast first, and then do the cast, or in the case of deopt, just do a virtual method call which is still much faster.

This is ~1.5% win for JIT compile speed.

Reviewed By: tekknolagi

Differential Revision: D36400173

fbshipit-source-id: 7e213697d60c291eb5d00f018a74a70fe9d1dfe6