Only support appending a single instruction at a time
Summary:
This removes the need for any of the special handling for the very rare cases where we wanted to emit multiple instructions at once, while making the code cleaner and more concise in the process.
This also eliminates the allocation of the line as a whole in effectively every case, unless a batch decref manages to produce more than 500 characters.

This is a ~1.5% build time win.

Reviewed By: tekknolagi

Differential Revision: D36433190

fbshipit-source-id: 3bea36d2b2d350d3864fc40be8df3c9cac3e1566