Add an API to touch memory used by profilers
Summary:
Some profilers need to walk the `code_rt->code->qualname` for jitted functions
on the call stack. The JIT rarely touches this memory and, as a result, the
OS may page it out. Out of process profilers (i.e. those that use eBPF) that
attempt to read the memory after it has been paged out will fail as the read
would cause a page fault. We can prevent the memory from being paged out as
aggressively by periodically reading it using the API included in this diff.

Reviewed By: czardoz

Differential Revision: D36580046

fbshipit-source-id: be0b9642bb7a3ed09e0cef3a3b9fac7d8ec61d37