Treat typed-dicts as dynamic
Summary:
Typed dicts are really just dicts at runtime, but treating them
as dicts in the compiler is a little tricky because returning plain
dicts with the correct keys is allowed.

In that sense, they are somewhat like named tuples. For now, let's treat
them as `dynamic`. We can always become stricter if this becomes
performance critical.

Reviewed By: carljm

Differential Revision: D36677143

fbshipit-source-id: 077f267c36822ec7f385c1ffc9a30de2afbe77c6