Re-apply D34926302 [cinder] Immortalize instances not discoverable through GC walk
Summary:
The diff was suspected to be one of causes of S270285 and was reverted to validate. There seems to be no evidence showing that the diff caused S270285 after the revert landed.

Re-applying the diff but with a knob to guard the code path to allow small roll-out and on/off switch.

Reviewed By: eduardo-elizondo

Differential Revision: D36550347

fbshipit-source-id: b99390d2c973b89283a712bb3560bd44db3b3dd9