Make semantics of cinderjit.force_compile() more useful
Summary:
Now returns `False` if the function was already compiled, `True` if compile succeeded, and raises a `RuntimeError` if compilation failed.

Also does not depend on the target function being "registered". If we can get a function that can compile without this, I see no reason why we shouldn't compile it.

Reviewed By: carljm

Differential Revision: D36801851

fbshipit-source-id: 51cfc1b8570c76e8f467279dacf1a9a0067b0100