Do not emit deferred compilation code for functions which can never compile
Summary:
We should not emit a call to `JITRT_CompileFunction()` if for some reason a function cannot compile with `_PyJIT_CompileFunction()`. Doing this will cause `JIRT_CompileFunction()` to waste time always taking the slow path.

I'm not sure if there's a case where we won't have a preloader but compilation may succeed later. In practice I haven't seen this happen with debug logging on a canary.

This change breaks Static Python enum returns in some cases because they are handled "better" in `JITRT_CompileFunction()` on the slow-path (and maybe on the fast path via JIT?). This bug is outside the scope of this diff and as discussed we may want to deprecated this area of enum support. So I've just disabled the failing tests.

Reviewed By: carljm

Differential Revision: D36730915

fbshipit-source-id: e4b1e795e81af4c6a22cc3e45226891bafbd6750