stop clearing the generic type cache when the classloader cache is cleared
Summary:
This is incorrect - as discussed in the Static Python slack channel, removing valid types from the generic instruction cache results in us creating multiple copies of valid types, and the equality check of Generic[T] == Generic[T] potentially failing as a result (since we have to mint a new type object, which won't compare equal pointer wise).

If we don't remove generic types from the cache, if you have an invalid type, we will leak some memory, but in the absence of building a mechanism to only clear out generic types with invalid parameters, this is necessary for correctness.

Reviewed By: carljm

Differential Revision: D36824755

fbshipit-source-id: 5260cb7679cf85efcb91283afc0da7643ed98054