Skip inlining for vectorcalled static functions with primitive args/return
Summary:
The inliner currently assumes all static JIT -> static JIT calls with primitive arguments and returns are made with `InvokeStaticFunction`. However, in some situations `HIRBuilder::emitInvokeFunction()` can emit a `VectorCall[Static]` along with a bunch of `PrimitiveBox` HIR nodes to converthe args into Python objects. The args are then Python objects instead of primitives as expected.

Max is working on a longer term fix that should handle this behavior correctly but for now should just not inline in this case.

Reviewed By: tekknolagi

Differential Revision: D36879865

fbshipit-source-id: ffb6b9a83793c7a7abaa9a286c70241d8bafe09c