Support storing the readonly operation mask in the shadow frame
Summary:
On 64-bit platforms all modern malloc implementations, including pymalloc, have a minimum guaranteed alignment of 16 bytes, so there is 1 remaining bit in the pointer that we can use to indicate whether we are currently in a readonly operation, and thus whether the mask can be safely read and inspected.

Since orig_data is unused and uninitialized unless a `PyFrameObject` has been materialized, we can safely use the space for the operation mask. For any frame that has been materialized, we continue to use the mask stored on the `PyFrameObject`.

Reviewed By: tiansihu

Differential Revision: D36466854

fbshipit-source-id: 8216498474458b88b915087a7d6e895ca25e342a