start clearing classloader caches when `sys.modules` is modified
Summary:
Context:

There's a subtle issue with lazy imports warmup and import cycles that's the impetus for this diff:
During the lazy imports warmup, we run module top-levels, potentially creating type objects, etc.
However, if we run into an import cycle, the partially initialized modules are silently dropped from
sys.modules. This has an unfortunate interaction with the classloader caches - the caches may have
been populated as a result of resolving type descrs with the type objects that are no longer used
anywhere.

To remove the stale data from the classloader cache, we start maintaining a dictionary of module -> keys to invalidate in case of a change.
Every time a given module is updated, we remove all descrs in this side dict, and clear that entry of the dict out.

Reviewed By: carljm, czardoz

Differential Revision: D36547418

fbshipit-source-id: 47a0c4c0dcda2d412d8b5fa522aff343d338ec0d