Tighten up deopt metadata
Summary:
Some simple changes to reduce the size of de-opt metadata, as it seems like we allocate a lot of it.  Changes small enums to be `char` sized and improves the layout of `DeoptMetadata`.  Also removes the `inline_depth` field as this can be inferred from the `frame_meta` field (unfortunately due to 8 byte alignment this doesn't save us memory, but would set us up to do so if we can get another field removed).

This reduces the size of `DeoptMetadata` from 104 bytes to 88, and `LiveValue` from 16 to 8.

Reviewed By: swtaarrs

Differential Revision: D36941657

fbshipit-source-id: 5971752ba8160a5ed99295cb7cf4ad4a39b32eeb