Warn on function args without type annotation
Summary:
Without further narrowing in the function body,
it is impossible to optimize attribute access or method calls
on function arguments without a type annotation,
since the compiler defaults back to `DYNAMIC`.

Raise a `PerfWarning` when this happens.

Also delete a `perf_warning` helper that overrides & duplicates the `GenericVisitor` implementation.

Reviewed By: tekknolagi

Differential Revision: D37152500

fbshipit-source-id: ab8b5efd50e3143d930e896ffa0701374d994f9f