Handle fmt in C++20 mode correctly
Summary:
With C++20 mode enabled, fmt uses consteval to check the validity of format strings, and produces an error if it's not checkable at compile time, so refactor various places to use `fmt::format_string` to forward the consteval'd string.
This also fixes one location where we were passing a name as a format string, and marks the couple of places where we have the format string as a configuration parameter as `fmt::runtime` so it doesn't error about not being able to evaluate it at compile time.

Reviewed By: swtaarrs

Differential Revision: D37185607

fbshipit-source-id: 40575380cbace4ed3ae9f0b22b5df029cb49775b