Make the full initialization of a few PyTypeObject and PyModuleDef explicit
Summary:
In C++17 mode, the mix of braced (via PyVarObject_HEAD_INIT) and dot initialized (like `.tp_name = "someName"`) generates a warning for newer versions of clang and gcc. Under C++20 mode this warning is suppressible for clang, but a hard, unsuppressable error under GCC.
To solve this, just use the same initialization form that the rest of python does.

Reviewed By: swtaarrs

Differential Revision: D37185727

fbshipit-source-id: f95899fe2577b234d2ae5337b49ef768380d3de4