Add cinder.get_and_clear_type_profiles_with_metadata()
Summary:
As a followup to the new profile data version in D36839855, export a
new function to managed code that contains split dict keys for profiled
types. I tried to structure it in a way that makes it easy to add more
type-specific metadata.

Also add `-X jit-profile-interp-period`, to allow setting the profile period
from the command line (there's already a way to set it from managed code).

Reviewed By: tekknolagi

Differential Revision: D37217536

fbshipit-source-id: 587306dd5fbc140a61e8baa72144f8472be5e497