Readonly for loop
Summary:
Implement readonly checks for the for loop.

### Semantics

For `for e1 in e2`, there are two sets of things to check.
Firstly, we need to check the operation of getting an iterator from e2. Since the iterator is immediately used (thus modified), the expected return type of `iter(e2)` is mutable. This means if a custom `__iter__` function returns a readonly iterator it cannot be used in a for loop. In general readonly iterators are not very useful.
Secondly, we need to check the operation of getting the next element. The expected input is mutable (enforced by the first check), and the expected output depends on the type of `e1`.

### Bytecode
Added readonly versions of the `GET_ITER` and `FOR_ITER` bytecodes
### Compiler
The `FOR_ITER` bytecode contains both a const and a relative jump. Therefore a new `emitWithBlock` method is added to the code generator which creates an instruction containing both the constant and the block for the jump. The block will be resolved to a constant relative jump distance, which is folded into the overall const tuple, during the graph flattening phase.
### Interpreter
Implemented readonly versions of `GET_ITER` and `FOR_ITER` which replicates the original logic but with readonly mask set ups and checks.
### Runtime
Added checks for `PyObject_GetIter` and `PyObject_SelfIter`. `PyObject_SelfIter` is also transitive w.r.t. readonly

### Misc
Also consolidated helpers for unit tests.

Reviewed By: Orvid

Differential Revision: D36458003

fbshipit-source-id: 399e7c7981a2d457905b15951c4973cba23bbfd8