Respect arg/return types for callables patched before compilation
Summary:
Currently, we do not account for the situation where a function
may be patched even before compilation happens. In such a situation, we
should look into the original callable to find the return type and arg
types.

In general, this is only required when we patch functions that
accept/return primitives, because in other cases, everything is a
PyObject.

Reviewed By: DinoV

Differential Revision: D37197138

fbshipit-source-id: 89d4eab00c8e25a05204a517adba41547fa5f544