Change CI config to use cinder310 project/VCS
Summary:
This is loosely based on D34061584, and will allow us to properly test
and land 3.10 diffs.

I added another exclusion list for tests that are imported but don't yet work
because they use Cinder features that we have to port to 3.10. There are also a
couple tests that don't specifically use Cinder features but need tweaks to
pass, and I'll create followup tasks for those.

Reviewed By: carljm

Differential Revision: D37524230

fbshipit-source-id: db7462a263987b091a7aca995624ef7a7d9e0440