no longer keep Final[T] annotations in the module table
Summary:
We keep track of final globals in the `named_finals` object of a ModuleTable, so we can eliminate the `Final[]` from the types. This change will allow us to lock down the ModuleTable.children table completely in an upcoming diff.

See the test for a case where we were previously printing Final[bool].

Reviewed By: wmeehan

Differential Revision: D37355774

fbshipit-source-id: f23eee7dc256b0b7a6c6ec4a6fa48dea3056f265