Skip unsupported distutils and grp tests
Summary:
From Matt's original commit for removing the negative test for non-existent group ids:
Unfortunately, on dev servers, `grp.getgrall()` returns only a subset
of gids that valid inputs to `grp.getgrgid()`. I'm not sure what sort
of shared library shenanigans we're playing on dev servers to make this
happen, but it seems unlikely that we will make a change that will break
this functionality.

Reviewed By: itamaro

Differential Revision: D37541268

fbshipit-source-id: 2a05d53ae509b53a8171c3ccc9b118322fbb7a8a