Port skeleton PyEntry_init, PyEntry_LazyInit, PyEntry_initnow
Summary:
Original diff: D22994776 by vladima, reviewed by dinoviehland

The original diff had some things in it that have already been removed
in `cinder/3.8` (e.g. `func_entry_po`), so I ported this `PyEntry_*` skeleton
from the current state of `cinder/3.8` rather than from the original diff.

Also in contrast to the original diff, I chose to just delete references to
the JIT and entrypoints that come later, rather than commenting them
out.

At this point this PyEntry stuff is pretty much just useless boilerplate,
but it will be used later by the JIT and fast-path entrypoints.

The current shape of this `PyEntry` stuff is not really upstreamable, but
it should eventually be replaced by use of whatever hooks we are able
to upstream in https://bugs.python.org/issue46897 and
https://bugs.python.org/issue46893

Did not port the change to force frame builtins to the default builtins. In
3.10, builtins are set on the `PyFunctionObject` at function creation time
and just copied to the frame, so there's no perf cost at frame creation.
The JIT does assume fixed builtins, but we can just have the JIT refuse
to compile a function with nonstandard builtins.

Reviewed By: itamaro

Differential Revision: D37542664

fbshipit-source-id: a5cb9b10bb210320daab6c7fc2e94b01fdaa949b