modify how annotations are passed to match the new expected format
Summary:
The 3.10 behavior of passing annotations is a bit different from the 3.8 version.

Instead of passing in a tuple of tuples, everything's flattened when `__future__.annotations` isn't present. When `from __future__ import annotations` is applied, the expectations is a `LOAD_CONST` with a tuple of the names. Let's change the Python compiler to work with this paradigm.

Reviewed By: itamaro

Differential Revision: D37581395

fbshipit-source-id: 989afa654c160b46ff8a8c7d036ddfd57985458e