adapt python compiler to the FINALLY opcode refactor
Summary:
The try..except..finally clause has been rewritten in https://github.com/python/cpython/pull/6641/files .

The gist is that instead of having the many _FINALLY opcodes, there's only SETUP_FINALLY and RERAISE, and the bodies of exception handling & with statements have been modified to no longer depend on the removed opcodes. Do a line-by-line port of the PR for the python compiler. There's a few other instructions that were added.

Some other interesting stuff surfaced while I was working on this:

1. In 3.10, the runtime uses instruction offsets instead of byte offsets, which means we no longer need to multiply offsets by two in the compiler, yay!
2. The comment for the `compiler_try_except` function seems wrong - the stack has 6 elements on it, and the bytecode we emit only eliminates 3 of the elements on the stack, but there's some other runtime magic that zeroes the stack when returning. I'm not sure what's going on there..

Can't run all the sbs code tests yet, but that's the goal.

Reviewed By: itamaro

Differential Revision: D37581393

fbshipit-source-id: 6ca6e38559f024e2e45320759d7315fbd85068ed