remove duplicate __debug__ handling
Summary:
We were handling `__debug__` both in the AST optimizer and again (redundantly)
in the compiler. Upstream handles it only in the AST optimizer; let's follow suit.

In order to make tests happy with this, we need to run flow graph optimizations
(and subsequent cleanups and normalizations) on the flow graph that we check
in various tests, so extract a `finalize()` method (and new `FINAL` state) to do this.

Also clean up some lingering references to obsolete forms of constant nodes.

Reviewed By: itamaro

Differential Revision: D37582016

fbshipit-source-id: 1126664a1434269964757b3c114a7c5b9fa5616d