fold constant tuples in compiler
Summary:
Previously, folding of tuples containing all constants into a single
`LOAD_CONST` of the entire tuple was done in the AstOptimizer,
but this has some limitations: e.g. tuples not directly present in the
AST but emitted synthetically in compilation (e.g. function argument
default value tuples) did not get folded.

In 3.10 the compiler itself also does constant tuple folding in flow
graph optimization, which handles these additional cases. Here we
add that to the py-compiler and add a corpus test to verify it.

C compiler code reference:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L7212

The upstream AST optimizer still also does the constant tuple folding,
so I've left it in our AST optimizer too. Previously our test for it in the
AST optimizer relied on comparisons of compiled bytecode with and
without AST optimizer; having the same optimization in the compiler
itself broke this test, so I rewrote the test to just directly test the
effect of the AST optimizer on an AST tree.

Code reference for tuple folding also still being present in upstream
AST optimizer:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/ast_opt.c#L313

Reviewed By: czardoz

Differential Revision: D37655985

fbshipit-source-id: 0ef566eee0ef9f098ba96747d5f5fec2b426efbe