simplify annotation building to match upstream
Summary:
We were going to some lengths in our compiler to ensure that when
`from __future__ import annotations` was active, we emitted a single
`LOAD_CONST` of a single tuple for annotations (since in this case
they are all string constants, so it's a fully constant tuple) rather than
a series of `LOAD_CONST` followed by `BUILD_TUPLE`.

The upstream compiler approach in 3.10 is simpler; it just emits the
longer `BUILD_TUPLE` version in either case initially, and relies on a
separate flow graph optimization pass (which I added earlier in this
diff stack) to reduce that to a single `LOAD_CONST` of the full tuple.

Failing to match this upstream behavior causes differences in the
contents of `co_consts` and the const indices in the bytecode, which
breaks our side-by-side bytecode verification. And the upstream
approach is simpler and nicer anyway.

Upstream compiler code for reference:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L2192

Reviewed By: czardoz

Differential Revision: D37667844

fbshipit-source-id: 5cbfacdd367dede66069867a692504892d395266