fix block traversal order in duplicate_exits_without_lineno
Summary:
The upstream compiler keeps all blocks in the compilation unit as a singly-linked list,
in the order they were allocated. This order occasionally matters to the emitted bytecode;
e.g. in `duplicate_exits_without_lineno` it affects the ordering of the duplicated blocks,
which changes jump targets. This difference is immaterial, but there's no way to
ignore just this particular difference and keep our side-by-side comparison tests.

We didn't have any equivalent of this allocation-order of blocks, and without it we
couldn't match the upstream-generated bytecode reliably. We were using execution
order, which usually is the same but not always. One case where they are different is
if/else statements: the "end" block is allocated before the "else" block.

This diffs adds block allocation-order tracking so we can match this behavior.
We could alternatively modify block allocation order in `visitIf`, but I think it's
better to keep the behavior matching upstream as closely as possible, so that
codegen methods can also stay close to upstream.

Reviewed By: czardoz

Differential Revision: D37703054

fbshipit-source-id: 8764054a0a277ac72eb7f1909901a372ea90a749