add missing copy_location() in AST optimizer
Summary:
Without this fix, the code in the added test would error in the compiler
due to missing `lineno` attribute on the synthetic `Tuple` AST node.

The call to `update_node` is useless without providing kwargs for attrs
to replace; it should just be a call to `copy_location` in this case where
the returned node is newly created.

Reviewed By: tekknolagi

Differential Revision: D37704607

fbshipit-source-id: b56f7fe7925da28a04413be5d60b4501cb426fec