Add Cinder-specific opcodes
Summary:
This diff ports all of the opcodes added to Cinder since the fork from upstream 3.8.

Keep opcode number the same as much as possible
while avoiding new upstream opcodes.

Reorder the `stack_effects` to match opcode ordering,
and move the readonly bytecode definitions to reflect that they are Cinder-internal.

For now, regenerate only `opcode.h` and not `opcode_targets.h`,
since we have not added targets for these opcodes to the interpreter yet.

Reviewed By: carljm

Differential Revision: D37829711

fbshipit-source-id: 1a1df0d6f2ef384dc4f176cfa3672b0731ba4e8a