set lineno for stmt nodes in visit()
Summary:
The upstream compiler sets the lineno for all statement nodes
generically in `compiler_visit_stmt`:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L3507

We were previously doing it at the top of every `visitFoo` method for
each type of statement node. Switch to the upstream approach
so that visit implementations line up more clearly, and for less
repetition.

This is doing the same thing for statements that D37870443 did for
expressions.

Reviewed By: itamaro

Differential Revision: D37894160

fbshipit-source-id: b8c5308c9dd1722af31d3ae3dc4e7dc567b589c3