improve typing and add None guard in flow graph optimizer
Summary:
Compilation of `Lib/idlelib/query.py` was failing due to `next_instr` being
`None` in the flow graph optimizer for `BUILD_TUPLE`. It's valid for that
to be `None`, we just weren't guarding for it correctly.

Improve/correct the type coverage so that this would have been caught
by the type checker, and add the needed `None` guard.

Reviewed By: itamaro

Differential Revision: D37896526

fbshipit-source-id: b7ed66d691c9c5b652eb0cd9792fd560017f1dd4