fix multi-jump-threading for first instr in block
Summary:
We were lazily checking truthiness of the return value of `dispatch_instr`
when we should have been checking if it was `None`. This meant that
if we optimized instruction offset 0 in a black, and returned 0 to mean
"check this instruction again," it was instead interpreted as "move on to
the next instruction" (which is what a return value of `None` should
mean.) This broke multiple-jump threading in the case where the jump
we are optimizing is the first instruction in its block.

Reviewed By: itamaro

Differential Revision: D37900189

fbshipit-source-id: 535bcde40f8296ed7c0eb875a2e7ae105ca8806a