start new block after visiting return
Summary:
Match upstream in starting a new block immediately after emitting a
`RETURN_VALUE`. See:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L3023

The story of why this makes any difference, and why the included test
demonstrates the difference, is more interesting. The symptom
observed in the included test was a `co_consts` that looked like
`("A docstring.", None, ())` upstream and `("A docstring.", ())` in our
compiler.

The reason is that at the end of a function, both we and upstream
check whether the current block ends with a return, and if it doesn't
we add `LOAD_CONST None; RETURN_VALUE` to it. Without this
automatic next-block at the end of `visitReturn`, we weren't adding that
(because the current block was still the returning one) but upstream
was (because the current block was the new, empty one.) So upstream
ended up with a redundant unreachable block at the end of the function
containing `LOAD_CONST None; RETURN_VALUE` which would later
be removed by DCE.

The next important factor is the `foo(**kw)`, which is one of the few
ways to get the compiler to emit a `BUILD_TUPLE` which will later
be optimized to a `LOAD_CONST (...)`. In most cases the compiler can
detect const tuples immediately and emit the `LOAD_CONST`, but in
the case of a `CALL_FUNCTION_EX` it emits `BUILD_TUPLE` for args
and relies on a later optimization pass to transform that to a const tuple
(in this case the empty tuple.) This means that there is a constant `()`
added to the const pool _after_ the redundant `None`, which means
that later, after DCE, when we try to trim unused consts from the _end_
of the const pool, we are not able to trim the now-unused `None`.

And the last important factor in the repro is the docstring, since
functions without a docstring always have `None` added to their
const pool anyway.

Our behavior in this specific case is better than upstream, but we want
to match upstream anyway, and there may be other cases where this
new block after return is more useful instead of counter-productive.

Reviewed By: itamaro

Differential Revision: D37902829

fbshipit-source-id: d3d9ed81d646af380f7738cc65f97f90736661ec