Port `Convert opcode.h into an X-macro`
Summary:
This brings over `fb5432c52c2`. The provided X-macro is used in a few
places in the JIT later in this stack.

I haven't run `make regen-opcode-targets` because we haven't added handlers
for the shadow bytecode and static python opcodes yet.

I didn't bump the magic number because we haven't changed the set of opcodes.

Reviewed By: swtaarrs

Differential Revision: D37893238

fbshipit-source-id: 0812074bf13c9db7b65e7801c8167063ab6a20e7