fix issue with lineno reverting to 0
Summary:
If the top-level node that we visit is an `expr`, which can be the case in a `lambda`, we
might revert back to the initial default lineno of `0` before emitting the final
`RETURN_VALUE`. This happens in the upstream compiler, too.

The difference is that we were unconditionally setting `prev_lineno` to each instr lineno
in our emit loop, even if it was 0. Upstream compiler doesn't do this; it leaves the
previous "current lineno" in place if an instruction has lineno 0. With this change, we
match that behavior and no longer emit a spurious extra tuple in `co_lines` for an
unchanged line number.

Reviewed By: itamaro

Differential Revision: D37946118

fbshipit-source-id: c6e54d696570500719ce17cc6deb0e89748c09bb