extend_blocks in reverse allocation order
Summary:
When iterating blocks to `extend_block` (copy exit blocks' opcodes into any block
that unconditionally jumps to that exit block), upstream does it in reverse block
allocation order, whereas we did it in forward order. This matters because in case of a
jump to a jump to an exit, forward order does not extend the first jump, but reverse
order will extend both (first the second jump gets extended into being an exit block in
its own right; now the first jump is jumping to an exit so it will get extended, too.)

See
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L7767
and
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/compile.c#L7121
-- iterating from `c->u->u_blocks` via `b->b_list` traverses the linked list of blocks in
reverse allocation order.

Reviewed By: itamaro

Differential Revision: D37975897

fbshipit-source-id: 9f5fd82ce8fc2f5e6aa603dd0306af1c34126190