move main logic of strict_modules.compileall to cinder
Summary:
Context: We used to have the logic of compiling a module in strict mode and writing the bytecode file live in other repositories. The way the integration is done by patching `py_compile.compile`. Instead of asking other repos to have the heavy lifting logic, let's move this into the loader so that we can update the compileall logic as the strict loader changes.

There's a slightly interesting design constraint here in that since we want to provide a patch to `py_compile.compile`, passing in repository-specific/loader settings is non-trivial. I chose to work around this by creating a higher order function that creates the patch candidate since it felt like a nice and easy pure option.

Reviewed By: carljm

Differential Revision: D37669135

fbshipit-source-id: 604782137ecec3a426ea3eedebd8e4dc305da1b2