completely control the .pyc bytecode reading/writing in the strict loader
Summary:
This is the first step for the bytecode invalidation work for proper dependency tracking in Static Python. The plan is to start tracking dependencies via tuples in the .pyc files of static modules (see https://fb.quip.com/awqXAtyCvl5e for details). To accomodate this, we need to take control of the reading of bytecodes by copying and adapting the get_code functionality in `Lib/importlib/_bootstrap_external.py`.

In regards of the implementation, where I plan to add the invalidation in the future is in the `compile_bytecode_from_pyc` function: You'll note that we've added a dummy dependencies tuple which is simply an empty tuple at the moment. The plan is to expand that, and add validation in the upcoming diffs before we read the code, and eagerly return None.

Reviewed By: carljm

Differential Revision: D37497155

fbshipit-source-id: 42d0faa825f5c69e4417156676d415bfbd29f003